refactor: move isModTeam to countingblacklist
